+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - `dc_receive_imf` refactorings #3154 #3156
 - add index to speedup deletion of expired ephemeral messages #3155
 - muted chats stay archived on new messages #3184
+- support `min_api` from Webxdc manifests #3206
 
 
 ### Fixes